backend/requirement.txt
sqlalchemy==2.0.44
python-dotenv==1.2.1
cryptography==46.0.3
snowflake-id==1.0.2
psycopg2-binary==2.9.11
fastapi==0.124.4
uvicorn==0.38.0